Avoid concurrent issues in MutableShellSession
When multiple MutableShellSessions are being used it's possible the
second one attempts to get the set log level while the first one is
still executing the code and depending on the order they finish we might
end up with a different log level than was originally set.

To avoid this lets move the machine.ShellSession into
utils.MutableShellSession and expect the root log level to be set in
utils.DEFAULT_ROOT_LOG_LEVEL. To ensure that we are setting it inside
the runperf.setup_logging.
